==> 1.R/02.Dados Consolidacao.Rmd <==
---
title: "Previsão ACs - Dados consolidação"
author: "João Isidio Freitas Martins"
date: "17/03/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

# Este documento requere a instalação dos seguintes pacotes:
# packages <- c('dplyr')

# Retire o comentário da próxima linha, basta remover o # (Ctrl+Shift+C), e execute o código.
# install.packages(packages)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

library(dplyr)
```

## Escopo

O trabalho consiste em estabelecer um método para a definição de modelos mais adequados para a previsão do número de atos de concentração dentro do CADE.

## Consolidação dos dados

O propósito deste código é o de unir às bases de atualização automática e não automática.
  - dados_api.rds
  - acs_cn.rds

Aqui serão criados todos os campos necessários para a avalição econométrica.

### Unindo as bases

Realiza-se a carga das bases:
```{r}
api<- readRDS(file = "3.output/1.data/dados_api.rds")
acs<- readRDS(file = "3.output/1.data/acs_cn.rds")
```

Função para acumulado móvel dos últimos 12 registros:
```{r}
mov<- 
  function(x){
    v<- NULL
    for(i in seq_along(x)){
      if(i <= 11){v[i]<- NA}
      else{v[i]<- sum(x[(i-11):i])}
      }
    v}

acs<-
acs %>%
  mutate(ac12c = mov(acs))
```

Só existe um campo comum, o de data. Realiza-se o Join.
```{r}
base<- dplyr::full_join(x = api, y = acs)
rm(acs, api, mov)
```

### Adicionando campos

Para controlar sazonalidades
```{r}
base$ano<- lubridate::year(x = base$date)
base$mes<- lubridate::month(x = base$date)
base$ano_mes<- paste0(base$ano, "-", formatC(x = base$mes, width = 2, flag = "0"))
```

### Tratamentos adicionais

Padronizando nomes
```{r}
nomes<-c(
  "data",
  "pibm",
  "pibac",
  "ibcm",
  "ibcsz",
  "slicm",
  "slica",
  "jmto",
  "jmpj",
  "cbef",
  "cbre",
  "cbnm",
  "ipca",
  "m4sld",
  "ibov",
  "embi",
  "pnad",
  "ac",
  "ac12c",
  # "lei",
  # "rs09",
  # "rs10",
  "ano",
  "mes",
  "anomes"
  )

names(base)<- nomes
```

Transformando o IPCA em índice (necessário para o log)
```{r}
# base$ipca<- (base$ipca/100)+1
```

Logaritmizando as variáveis
```{r}
# variaveis<- nomes[!nomes %in% c("data", "ac", "ac12c", "ano", "mes", "anomes")]
# 
# base[,variaveis]<- sapply(X = base[,variaveis], FUN = log)
```

Salvando o RDS
```{r}
saveRDS(object = base, file = "3.output/1.data/base.rds")
```

Removendo lixo:
```{r}
rm(list = ls())
```

FIM
